fix:fragmentation implementation completed on 21/08/2024

+++ app/src/main/java/com/example/mysoreprintersproject/responses/ExecutiveDashboard.kt
@@ -23,6 +23,6 @@ data class MonthlyReport (
 data class AttendanceGraph (
 
     @SerializedName("month"                 ) var month                : String? = null,
-    @SerializedName("attendance_percentage" ) var attendancePercentage : Int?    = null
+    @SerializedName("attendance_percentage" ) var attendancePercentage : Double?    = null
 
 )
